feat: 优化常量
