+++ src/components/WordMarkLayout/context.ts
@@ -1,5 +1,5 @@
 import { createContext } from 'react';
-import { IWordMarkConfig } from '@/isomorphic/word-mark';
+import { IWordMarkConfig } from '@/isomorphic/constant/wordMark';
 
 export interface IWordMarkContext extends IWordMarkConfig {}
 export const WordMarkContext = createContext<IWordMarkContext>({} as IWordMarkContext);
