--- src/context/account-context.ts
@@ -1,14 +1,0 @@
-import { createContext } from 'react';
-import { IUser } from '@/core/account';
-
-export interface IAccountContext {
-  user: IUser;
-  onLogout: () => void;
-}
-
-export const AccountContext = createContext<IAccountContext>({
-  user: {} as IUser,
-  onLogout: () => {
-    //
-  },
-});
