--- src/context/word-mark-context.ts
@@ -1,6 +1,0 @@
-import { IWordMarkConfig } from '@/isomorphic/word-mark';
-import { createContext } from 'react';
-export interface IWordMarkContext extends IWordMarkConfig {
-  destroyWordMark: () => void;
-}
-export const WordMarkContext = createContext<IWordMarkContext>({} as IWordMarkContext);
