+++ src/isomorphic/link-helper.ts
@@ -17,6 +17,8 @@ const LinkHelper = {
   feedback: 'https://www.yuque.com/feedbacks/new',
   joinGroup:
     'https://www.yuque.com/yuque/yuque-browser-extension/welcome#BQrrd',
+  settingPage: chrome.runtime.getURL('setting.html'),
+  wordMarkSettingPage: `${chrome.runtime.getURL('setting.html')}?page=wordMark`,
 };
 
 export default LinkHelper;
