+++ src/pages/inject/WordMark/Inner/OperateMenu.tsx
@@ -4,7 +4,7 @@ import { PushpinOutlined, PushpinFilled } from '@ant-design/icons';
 import classnames from 'classnames';
 import { backgroundBridge } from '@/core/bridge/background';
 import { WordMarkOptionTypeEnum } from '@/isomorphic/constants';
-import { WordMarkConfigKey } from '@/isomorphic/word-mark';
+import { WordMarkConfigKey } from '@/isomorphic/constant/wordMark';
 import { useWordMarkContext } from '@/components/WordMarkLayout/useWordMarkContext';
 import DragList from '@/components/DragList';
 import { ToolbarItem, toolbars as defaultToolbars } from '../constants';
