+++ src/pages/inject/WordMark/Inner/index.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState } from 'react';
 import { Popover } from 'antd';
 import Icon, { CloseOutlined } from '@ant-design/icons';
 import { WordMarkOptionTypeEnum } from '@/isomorphic/constants';
-import { WordMarkConfigKey } from '@/isomorphic/word-mark';
+import { WordMarkConfigKey } from '@/isomorphic/constant/wordMark';
 import { backgroundBridge } from '@/core/bridge/background';
 import { useWordMarkContext } from '@/components/WordMarkLayout/useWordMarkContext';
 import YuqueLogo from '@/assets/svg/yuque-logo.svg';
