--- src/pages/inject/WordMark/util.ts
@@ -1,3 +1,0 @@
-export const getPageUrl = () => {
-  return `${window.location.origin}${window.location.pathname}`;
-};
