--- src/pages/inject/browser-script.ts
@@ -1,3 +1,0 @@
-import { initI18N } from '@/isomorphic/i18n';
-
-initI18N();
