+++ src/pages/setting/help/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { __i18n } from '@/isomorphic/i18n';
-import LinkHelper from '@/core/link-helper';
+import LinkHelper from '@/isomorphic/link-helper';
 import styles from './index.module.less';
 
 function Help() {
