+++ src/pages/setting/wordMark/index.module.less
@@ -1,0 +1,41 @@
+@import '~@/styles/parameters.less';
+
+.configWrapper {
+  max-width: 668px;
+
+  .card {
+    padding: 24px 0;
+    border-bottom: 1px solid @border-light-color;
+    color: @text-primary;
+
+    .title {
+      font-size: @font-size-lg;
+      line-height: 24px;
+      margin-bottom: 16px;
+    }
+
+    .body {
+      display: flex;
+      flex-direction: column;
+      gap: 12px;
+
+      .configItem {
+        display: flex;
+        justify-content: space-between;
+
+        .desc {
+          font-size: @font-size;
+        }
+      }
+    }
+  }
+
+  .card:first-child {
+    padding-top: 0;
+  }
+
+  .card:last-child {
+    border-bottom: 1px solid transparent;
+  }
+}
+
