+++ src/pages/sidePanel/app.module.less
@@ -6,6 +6,6 @@
 }
 
 .sidePanelWrapper {
-  width: 418px;
+  width: 100%;
   height: 100%;
 }
